pyproject.toml
[tool.poetry]
name = "dev-days-flask"
version = "0.1.0"
description = ""
authors = ["sheryak_n <nathanraysheryak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
marshmallow = "^3.17.0"
marshmallow-enum = "^1.5.1"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
